src/models/smolvlm2/requirements.txt
# SmolVLM2 MLX Dependencies for Apple Silicon
# Optimized for MacBook Air/Pro with Apple Neural Engine

# Core MLX framework for Apple Silicon optimization
mlx>=0.12.0
mlx-lm>=0.8.0

# MLX-VLM for SmolVLM2 support
git+https://github.com/pcuenca/mlx-vlm.git@smolvlm

# Core dependencies
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.49.0
tokenizers>=0.15.0

# Image and video processing
Pillow>=10.0.0
opencv-python>=4.8.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# API and web server
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0

# Data handling
numpy>=1.24.0
scipy>=1.11.0

# Configuration and utilities
pydantic>=2.5.0
python-multipart>=0.0.6
aiofiles>=23.2.0

# Development and debugging (optional)
jupyter>=1.0.0
matplotlib>=3.7.0
tqdm>=4.66.0

# Apple Silicon optimizations
accelerate>=0.24.0

# Additional utilities for video processing
ffmpeg-python>=0.2.0
moviepy>=1.0.3 